+++ tetris/constants.py
@@ -1,0 +1,4 @@
+HEIGHT = 10
+WIDTH = 23
+TIME_INTERVAL = 0.1
+VERSION = 'v1.0.0'
