+++ tetris/gui/__init__.py
@@ -1,0 +1,1 @@
+from tetris.gui.ui import UI
