+++ tetris/gui/ui.py
@@ -1,93 +1,14 @@
-import os
-import sys
+import os, sys
 import curses
 import textwrap
 from contextlib import contextmanager
 
-from tetris.game import *
-from tetris.welcome import *
-from tetris.setup import *
+from tetris.core import Game, Pos, Color
+from tetris.gui.welcome import welcome_message
+from tetris.gui.objects import *
 from tetris.constants import TIME_INTERVAL, VERSION
 
 from tetris import logger
-
-def clean_and_split(string):
-    lines = textwrap.dedent(string.strip()).splitlines()
-    return [l.strip() for l in lines if l.strip()]
-
-class Box:
-    def __init__(self, *args, min_height=3, min_width=34):
-        self.min_height = min_height
-        self.min_width = min_width
-        self.lines = []
-
-        for i, text in enumerate(args):
-            bottom = 0 if i == len(args) - 1 else 1
-            self.add_text(text, bottom=bottom)
-
-    def add_text(self, text, top=0, bottom=0, color=None):
-        pad = (('', color),)
-        if not text:
-            self.lines.extend(pad)
-        else:
-            lines = clean_and_split(text)
-            self.lines.extend(pad * top)
-            self.lines.extend(zip(lines, (color,) * len(lines)))
-            self.lines.extend(pad * bottom)
-            longest_line = max(len(line) for line in lines)
-            self.min_width = max(self.min_width, longest_line)
-            return self # for fluid interface
-
-    def render(self, ui, pos=None):
-        actual_height = len(self.lines)
-        pad_amount = max(0, self.min_height - actual_height)
-        box_height = max(self.min_height, actual_height)
-        pos = pos or Pos(
-            ui.height // 2 - box_height // 2,
-            ui.width // 2 - self.min_width // 2 - 1)
-
-        offset = 0
-        def add_line(line, color=None):
-            nonlocal offset
-            color = color or ui.board_color
-            ui.stdscr.addstr(pos.y + offset, pos.x, line, color)
-            offset += 1
-
-        def add_lines(lines, colored=False):
-            for line in lines:
-                color = None
-                if colored:
-                    line, color = line
-                add_line(line, color)
-
-        center = '-'*self.min_width
-        pad = '|{}|'.format(' '*self.min_width)
-        padding = (pad,) * (pad_amount // 2)
-        lines = (('|{}|'.format(l.center(self.min_width)), c) for l,c in self.lines)
-
-        add_line('┌{}┐'.format(center))
-        add_lines((pad,) * (pad_amount // 2))
-        add_lines(lines, colored=True)
-        add_lines((pad,) * (pad_amount // 2 + pad_amount % 2))
-        add_line('└{}┘'.format(center))
-
-class Dialog(Box):
-    def __init__(self, *args, min_height=3, min_width=34, **kwargs):
-        super().__init__(*args, min_height=min_height, min_width=min_width)
-        self.add_keys(**kwargs)
-
-    def add_keys(self, **kwargs):
-        self.keys = kwargs.keys()
-        self.add_text(None) # ne
-        for key, msg in kwargs.items():
-            self.add_text("Press `{}` to {}".format(key, msg))
-        return self
-
-    @contextmanager
-    def response(self, ui):
-        self.render(ui)
-        with ui.get_key(self.keys, delay=True) as key:
-            yield key
 
 class UI:
     def __init__(self):
@@ -135,13 +56,13 @@ class UI:
     def setup_colors(self):
         curses.start_color()
         self.has_colors = curses.has_colors()
-        for color in setup.Color:
+        for color in Color:
             curses.init_pair(color.value, color.value, color.value)
         curses.init_pair(10, curses.COLOR_WHITE, curses.COLOR_BLACK)
         self.board_color = curses.color_pair(10)
 
     def doWelcome(self):
-        self.stdscr.addstr(0, 0, welcomeMessage[0])
+        self.stdscr.addstr(0, 0, welcome_message[0])
         blink = True
         animate_counter = 0
         refresh_counter = 0
@@ -150,8 +71,8 @@ class UI:
                 if key == ' ':
                     return
             if refresh_counter % 10 == 0:
-                self.stdscr.addstr(0, 0, welcomeMessage[animate_counter])
-                animate_counter = (animate_counter + 1) % len(welcomeMessage)
+                self.stdscr.addstr(0, 0, welcome_message[animate_counter])
+                animate_counter = (animate_counter + 1) % len(welcome_message)
                 if blink:
                     Box().render(self)
                 else:
@@ -250,11 +171,12 @@ class UI:
         self.displayBoard()
         left = 28
         # score
-        Box("level: {}".format(self.g.level),
-            "lines: {}".format(self.g.cleared_lines),
+        Box("level: {}".format(self.g.level), "lines: {}".format(self.g.cleared_lines),
             min_width=18
         ).render(self, pos=Pos(11, left))
-        Box("score", '{} pts'.format(self.g.score), min_width=18).render(self, pos=Pos(19, left))
+        Box("score", '{} pts'.format(self.g.score),
+            min_width=18
+        ).render(self, pos=Pos(19, left))
         # next piece box
         Box(min_width=18, min_height=6).render(self, pos=Pos(2, left))
         nextPieceLines = self.g.next_piece.to_lines()
